Gate prost and serde dependencies behind features
