+++ src/storage/mod.rs
@@ -1,3 +1,7 @@
+#[cfg(feature = "serde")]
+mod codec_json;
+#[cfg(feature = "prost")]
+mod codec_prost;
 mod use_storage;
 
 pub use use_storage::*;
